Split terratorch IO Processor plugins tests from generic ones
